renamed objects -> mapIcons
