+++ src/components/map/layers/mapIcons/NodeMarker.tsx
