+++ src/components/map/layers/mapIcons/StartMarker.tsx
