+++ src/components/map/layers/mapIcons/nodeMarker.scss
